Removed console input
Won't be using console input anymore, the line function was changed to another file, each type of drawing will be inside a different file.
