+++ header/line.h
@@ -1,0 +1,9 @@
+#ifndef LINE_H
+#define LINE_H
+
+#include "structs.h"
+#include "base_functions.h"
+
+struct Image drawLine(struct Point p0, struct Point p1);
+
+#endif //LINE_H
